console: Refactor device-add view
